Adds estomators

+++ src/main/java/membership/Membership.java
@@ -2,6 +2,6 @@ package membership;
 
 public interface Membership {
     void init(int n, double p);
-    void insert(int key);
-    boolean contains(int key);
+    void insert(long key);
+    boolean contains(long key);
 }
